This is a horrid way to manage source control. Sorry.
Right now, the terrain is drawn using preset ROCKINESS and SMOOTHNESS multipliers in TPTerrainGenerator. Each chunk has a TPTerrainGenerator object which is initialized with the map seed and map size, and is then used to determine heights for given points. I used a 16 p-km (my own measurement, the perfect kilometer. 1024 meters) map with this method and the performance was acceptable with the view range set to 128. Not great, but pretty good.

Also, I've put together a little library to help with drawing and managing opengl, called atmosphere. I'm not using in the actual terrain drawing code because it's written in C and c structs cannot be easily put in and taken out of nsarray's. perhaps I'll look into changing systems in the future, but for now, it's straight opengl.

There is a rudimentary console system that sort of allows me to print out values to the opengl context. I really want to play around with that and make it a little easier to use.

That's it! I'll attempt to do version control better from here on. I think this is my second commit :-).